bugfix(TUP-24695) Refactor
