+++ main/plugins/org.talend.designer.maven.aether/.classpath
@@ -1,9 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <classpath>
-	<classpathentry exported="true" kind="lib" path="lib/commons-io.jar"/>
-	<classpathentry exported="true" kind="lib" path="lib/jsoup.jar"/>
-	<classpathentry exported="true" kind="lib" path="lib/slf4j-simple.jar"/>
 	<classpathentry exported="true" kind="lib" path="lib/wagon-file.jar"/>
+	<classpathentry exported="true" kind="lib" path="lib/jsoup.jar"/>
 	<classpathentry exported="true" kind="lib" path="lib/wagon-http-shared.jar"/>
 	<classpathentry exported="true" kind="lib" path="lib/wagon-http.jar"/>
 	<classpathentry exported="true" kind="lib" path="lib/commons-codec.jar"/>
@@ -15,8 +13,6 @@
 	<classpathentry exported="true" kind="lib" path="lib/maven-resolver-impl.jar"/>
 	<classpathentry exported="true" kind="lib" path="lib/maven-resolver-spi.jar"/>
 	<classpathentry exported="true" kind="lib" path="lib/maven-resolver-transport-classpath.jar"/>
-	<classpathentry exported="true" kind="lib" path="lib/maven-resolver-transport-file.jar"/>
-	<classpathentry exported="true" kind="lib" path="lib/maven-resolver-transport-http.jar"/>
 	<classpathentry exported="true" kind="lib" path="lib/maven-resolver-transport-wagon.jar"/>
 	<classpathentry exported="true" kind="lib" path="lib/maven-resolver-util.jar"/>
 	<classpathentry exported="true" kind="lib" path="lib/plexus-utils.jar"/>
