+++ main/plugins/org.talend.designer.maven.aether/META-INF/MANIFEST.MF
@@ -26,8 +26,6 @@ Bundle-ClassPath: .,
  lib/maven-resolver-impl.jar,
  lib/maven-resolver-spi.jar,
  lib/maven-resolver-transport-classpath.jar,
- lib/maven-resolver-transport-file.jar,
- lib/maven-resolver-transport-http.jar,
  lib/maven-resolver-transport-wagon.jar,
  lib/maven-resolver-util.jar,
  lib/plexus-utils.jar,
@@ -35,7 +33,8 @@ Bundle-ClassPath: .,
  lib/wagon-provider-api.jar,
  lib/jsoup.jar,
  lib/wagon-http-shared.jar,
- lib/wagon-http.jar
+ lib/wagon-http.jar,
+ lib/wagon-file.jar
 Export-Package: org.talend.designer.maven.aether,
  org.talend.designer.maven.aether.comparator,
  org.talend.designer.maven.aether.node,
