+++ main/plugins/org.talend.designer.maven.aether/build.properties
@@ -5,4 +5,5 @@ bin.includes = META-INF/,\
                lib/,\
                lib/jsoup.jar,\
                lib/wagon-http-shared.jar,\
-               lib/wagon-http.jar
+               lib/wagon-http.jar,\
+               lib/wagon-file.jar
