+++ main/plugins/org.talend.designer.maven.aether/src/main/java/org/talend/designer/maven/aether/RepositorySystemFactory.java
@@ -17,33 +17,20 @@ import java.util.HashMap;
 import java.util.Map;
 
 import org.apache.maven.repository.internal.MavenRepositorySystemUtils;
-import org.apache.maven.wagon.Wagon;
-import org.apache.maven.wagon.providers.file.FileWagon;
-import org.apache.maven.wagon.providers.http.HttpWagon;
-import org.codehaus.plexus.DefaultPlexusContainer;
-import org.codehaus.plexus.PlexusContainer;
 import org.codehaus.plexus.PlexusContainerException;
 import org.eclipse.aether.DefaultRepositorySystemSession;
 import org.eclipse.aether.RepositorySystem;
 import org.eclipse.aether.artifact.Artifact;
 import org.eclipse.aether.artifact.DefaultArtifact;
-import org.eclipse.aether.connector.basic.BasicRepositoryConnectorFactory;
 import org.eclipse.aether.deployment.DeployRequest;
-import org.eclipse.aether.impl.DefaultServiceLocator;
-import org.eclipse.aether.internal.transport.wagon.PlexusWagonConfigurator;
-import org.eclipse.aether.internal.transport.wagon.PlexusWagonProvider;
 import org.eclipse.aether.repository.Authentication;
 import org.eclipse.aether.repository.LocalRepository;
 import org.eclipse.aether.repository.RemoteRepository;
-import org.eclipse.aether.spi.connector.RepositoryConnectorFactory;
-import org.eclipse.aether.spi.connector.transport.TransporterFactory;
-import org.eclipse.aether.transport.file.FileTransporterFactory;
-import org.eclipse.aether.transport.http.HttpTransporterFactory;
-import org.eclipse.aether.transport.wagon.WagonTransporterFactory;
 import org.eclipse.aether.util.artifact.SubArtifact;
 import org.eclipse.aether.util.listener.ChainedRepositoryListener;
 import org.eclipse.aether.util.listener.ChainedTransferListener;
 import org.eclipse.aether.util.repository.AuthenticationBuilder;
+import org.talend.designer.maven.aether.util.MavenLibraryResolverProvider;
 
 /**
  * created by wchen on Aug 10, 2017 Detailled comment
@@ -53,32 +40,6 @@ public class RepositorySystemFactory {
 
     private static Map<LocalRepository, DefaultRepositorySystemSession> sessions = new HashMap<LocalRepository, DefaultRepositorySystemSession>();
 
-    private static RepositorySystem newRepositorySystem() throws PlexusContainerException {
-        DefaultServiceLocator locator = MavenRepositorySystemUtils.newServiceLocator();
-        locator.addService(RepositoryConnectorFactory.class, BasicRepositoryConnectorFactory.class);
-        // TUP-24695 change to wagon transporters
-        locator.addService(TransporterFactory.class, WagonTransporterFactory.class);
-
-        PlexusContainer pc = new DefaultPlexusContainer();
-
-        pc.addComponent(new HttpWagon(), Wagon.class, "http");
-        pc.addComponent(new FileWagon(), Wagon.class, "file");
-
-        WagonTransporterFactory tf = (WagonTransporterFactory) locator.getService(TransporterFactory.class);
-        tf.setWagonConfigurator(new PlexusWagonConfigurator(pc));
-        tf.setWagonProvider(new PlexusWagonProvider(pc));
-
-        locator.setErrorHandler(new DefaultServiceLocator.ErrorHandler() {
-
-            @Override
-            public void serviceCreationFailed(Class<?> type, Class<?> impl, Throwable exception) {
-                exception.printStackTrace();
-            }
-        });
-
-        return locator.getService(RepositorySystem.class);
-    }
-
     private static DefaultRepositorySystemSession newRepositorySystemSession(String localRepositoryPath)
             throws PlexusContainerException {
         LocalRepository localRepo = new LocalRepository(localRepositoryPath);
@@ -87,7 +48,8 @@ public class RepositorySystemFactory {
             repositorySystemSession = MavenRepositorySystemUtils.newSession();
             repositorySystemSession
                     .setLocalRepositoryManager(
-                            newRepositorySystem().newLocalRepositoryManager(repositorySystemSession, localRepo));
+                            MavenLibraryResolverProvider.newRepositorySystem().newLocalRepositoryManager(repositorySystemSession,
+                                    localRepo));
             repositorySystemSession.setTransferListener(new ChainedTransferListener());
             repositorySystemSession.setRepositoryListener(new ChainedRepositoryListener());
         }
@@ -99,7 +61,7 @@ public class RepositorySystemFactory {
             String userName, String password, String groupId, String artifactId, String classifier, String extension,
             String version) throws Exception {
         DefaultRepositorySystemSession session = null;
-        RepositorySystem system = newRepositorySystem();
+        RepositorySystem system = MavenLibraryResolverProvider.newRepositorySystem();
         session = newRepositorySystemSession(localRepository);
 
         DeployRequest deployRequest = new DeployRequest();
