+++ main/plugins/org.talend.designer.maven.aether/src/main/java/org/talend/designer/maven/aether/util/MavenLibraryResolverProvider.java
@@ -18,6 +18,12 @@ import java.util.Map;
 import org.apache.maven.model.License;
 import org.apache.maven.model.Model;
 import org.apache.maven.repository.internal.MavenRepositorySystemUtils;
+import org.apache.maven.wagon.Wagon;
+import org.apache.maven.wagon.providers.file.FileWagon;
+import org.apache.maven.wagon.providers.http.HttpWagon;
+import org.codehaus.plexus.DefaultPlexusContainer;
+import org.codehaus.plexus.PlexusContainer;
+import org.codehaus.plexus.PlexusContainerException;
 import org.eclipse.aether.DefaultRepositorySystemSession;
 import org.eclipse.aether.RepositorySystem;
 import org.eclipse.aether.RepositorySystemSession;
@@ -25,6 +31,8 @@ import org.eclipse.aether.artifact.Artifact;
 import org.eclipse.aether.artifact.DefaultArtifact;
 import org.eclipse.aether.connector.basic.BasicRepositoryConnectorFactory;
 import org.eclipse.aether.impl.DefaultServiceLocator;
+import org.eclipse.aether.internal.transport.wagon.PlexusWagonConfigurator;
+import org.eclipse.aether.internal.transport.wagon.PlexusWagonProvider;
 import org.eclipse.aether.repository.Authentication;
 import org.eclipse.aether.repository.LocalRepository;
 import org.eclipse.aether.repository.RemoteRepository;
@@ -32,8 +40,7 @@ import org.eclipse.aether.resolution.ArtifactRequest;
 import org.eclipse.aether.resolution.ArtifactResult;
 import org.eclipse.aether.spi.connector.RepositoryConnectorFactory;
 import org.eclipse.aether.spi.connector.transport.TransporterFactory;
-import org.eclipse.aether.transport.file.FileTransporterFactory;
-import org.eclipse.aether.transport.http.HttpTransporterFactory;
+import org.eclipse.aether.transport.wagon.WagonTransporterFactory;
 import org.eclipse.aether.util.repository.AuthenticationBuilder;
 import org.eclipse.m2e.core.MavenPlugin;
 import org.talend.core.nexus.ArtifactRepositoryBean;
@@ -70,7 +77,7 @@ public class MavenLibraryResolverProvider {
         return instance;
     }
 
-    private MavenLibraryResolverProvider() {
+    private MavenLibraryResolverProvider() throws PlexusContainerException {
         defaultRepoSystem = newRepositorySystem();
         defaultRepoSystemSession = newSession(defaultRepoSystem, getLocalMVNRepository());
         ArtifactRepositoryBean talendServer = TalendLibsServerManager.getInstance().getTalentArtifactServer();
@@ -146,12 +153,28 @@ public class MavenLibraryResolverProvider {
         return repository;
     }
 
-    private RepositorySystem newRepositorySystem() {
+    public static RepositorySystem newRepositorySystem() throws PlexusContainerException {
         DefaultServiceLocator locator = MavenRepositorySystemUtils.newServiceLocator();
         locator.addService(RepositoryConnectorFactory.class, BasicRepositoryConnectorFactory.class);
-        locator.addService(TransporterFactory.class, FileTransporterFactory.class);
-        locator.addService(TransporterFactory.class, HttpTransporterFactory.class);
+        // TUP-24695 change to wagon transporters
+        locator.addService(TransporterFactory.class, WagonTransporterFactory.class);
 
+        PlexusContainer pc = new DefaultPlexusContainer();
+
+        pc.addComponent(new HttpWagon(), Wagon.class, "http");
+        pc.addComponent(new FileWagon(), Wagon.class, "file");
+
+        WagonTransporterFactory tf = (WagonTransporterFactory) locator.getService(TransporterFactory.class);
+        tf.setWagonConfigurator(new PlexusWagonConfigurator(pc));
+        tf.setWagonProvider(new PlexusWagonProvider(pc));
+
+        locator.setErrorHandler(new DefaultServiceLocator.ErrorHandler() {
+
+            @Override
+            public void serviceCreationFailed(Class<?> type, Class<?> impl, Throwable exception) {
+                exception.printStackTrace();
+            }
+        });
         return locator.getService(RepositorySystem.class);
     }
